requirements.txt
alembic==1.3.1
Click==7.0
ecdsa==0.14.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
future==0.18.2
gunicorn==20.0.3
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
pycryptodome==3.3.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose-cryptodome==1.3.2
six==1.13.0
SQLAlchemy==1.3.11
Werkzeug==0.16.0
python-dotenv
